requirements.txt
Django==5.2.4
djangorestframework==3.16.0
Pillow==11.3.0
psycopg2-binary==2.9.10
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-leaflet==0.32.0
django-storages==1.14.6
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.9.0 

WeasyPrint>=61.0 